Rename `classCustomGraphicsItems.py` -> `graphics_items.py`
